refactor(lesser-peer): reorganize monitor code into separate namespace
- Move monitor-specific code to catl::peer::monitor namespace
- Reorganize directory structure:
  - Monitor headers: includes/catl/peer/monitor/
  - Monitor sources: src/monitor/
- Keep generic peer functionality in catl::peer namespace:
  - peer-connection, crypto-utils, packet-names, types
- Update all includes and namespace declarations
- Update CMakeLists.txt for new file locations
- Add forward declaration and update friend access for peer_connection

This separation allows the generic peer library to be reused for other
applications while keeping monitor-specific code isolated.
